DMP-181 update after review

+++ src/main/java/uk/gov/hmcts/darts/courthouse/CourthouseRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 @Repository
 public interface CourthouseRepository extends JpaRepository<Courthouse, Integer> {
 
-    Optional<Courthouse> findByCode(Short code);
+    Optional<Courthouse> findByCode(int code);
 
     Optional<Courthouse> findByCourthouseName(String name);
 }
